Move logging configuration into `alertmanager` package
