+++ pkg/component/monitoring/alertmanager/logging.go
@@ -1,4 +1,4 @@
-// Copyright 2023 SAP SE or an SAP affiliate company. All rights reserved. This file is licensed under the Apache Software License, v. 2 except as noted otherwise in the LICENSE file
+// Copyright 2024 SAP SE or an SAP affiliate company. All rights reserved. This file is licensed under the Apache Software License, v. 2 except as noted otherwise in the LICENSE file
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,11 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package monitoring
+package alertmanager
 
 import (
-	"fmt"
-
 	fluentbitv1alpha2 "github.com/fluent/fluent-operator/v2/apis/fluentbit/v1alpha2"
 	fluentbitv1alpha2filter "github.com/fluent/fluent-operator/v2/apis/fluentbit/v1alpha2/plugins/filter"
 	fluentbitv1alpha2parser "github.com/fluent/fluent-operator/v2/apis/fluentbit/v1alpha2/plugins/parser"
@@ -27,11 +25,7 @@ import (
 	"github.com/gardener/gardener/pkg/component"
 )
 
-var (
-	prometheusName = "prometheus"
-)
-
-// CentralLoggingConfiguration returns a fluent-bit parser and filter for the monitoring logs.
+// CentralLoggingConfiguration returns a fluent-bit parser and filter for the alertmanager logs.
 func CentralLoggingConfiguration() (component.CentralLoggingConfig, error) {
 	return component.CentralLoggingConfig{Filters: generateClusterFilters(), Parsers: generateClusterParsers()}, nil
 }
@@ -40,34 +34,16 @@ func generateClusterFilters() []*fluentbitv1alpha2.ClusterFilter {
 	return []*fluentbitv1alpha2.ClusterFilter{
 		{
 			ObjectMeta: metav1.ObjectMeta{
-				Name:   fmt.Sprintf("%s--%s", prometheusName, prometheusName),
+				Name:   "alertmanager",
 				Labels: map[string]string{v1beta1constants.LabelKeyCustomLoggingResource: v1beta1constants.LabelValueCustomLoggingResource},
 			},
 			Spec: fluentbitv1alpha2.FilterSpec{
-				Match: fmt.Sprintf("kubernetes.*%s*%s*", prometheusName, prometheusName),
+				Match: "kubernetes.*alertmanager*alertmanager*",
 				FilterItems: []fluentbitv1alpha2.FilterItem{
 					{
 						Parser: &fluentbitv1alpha2filter.Parser{
 							KeyName:     "log",
-							Parser:      prometheusName + "-parser",
-							ReserveData: ptr.To(true),
-						},
-					},
-				},
-			},
-		},
-		{
-			ObjectMeta: metav1.ObjectMeta{
-				Name:   prometheusName + "--blackbox-exporter",
-				Labels: map[string]string{v1beta1constants.LabelKeyCustomLoggingResource: v1beta1constants.LabelValueCustomLoggingResource},
-			},
-			Spec: fluentbitv1alpha2.FilterSpec{
-				Match: fmt.Sprintf("kubernetes.*%s*blackbox-exporter*", prometheusName),
-				FilterItems: []fluentbitv1alpha2.FilterItem{
-					{
-						Parser: &fluentbitv1alpha2filter.Parser{
-							KeyName:     "log",
-							Parser:      prometheusName + "-parser",
+							Parser:      "alertmanager-parser",
 							ReserveData: ptr.To(true),
 						},
 					},
@@ -81,14 +57,14 @@ func generateClusterParsers() []*fluentbitv1alpha2.ClusterParser {
 	return []*fluentbitv1alpha2.ClusterParser{
 		{
 			ObjectMeta: metav1.ObjectMeta{
-				Name:   prometheusName + "-parser",
+				Name:   "alertmanager-parser",
 				Labels: map[string]string{v1beta1constants.LabelKeyCustomLoggingResource: v1beta1constants.LabelValueCustomLoggingResource},
 			},
 			Spec: fluentbitv1alpha2.ParserSpec{
 				Regex: &fluentbitv1alpha2parser.Regex{
-					Regex:      "^ts=(?<time>\\d{4}-\\d{2}-\\d{2}[Tt]{1}\\d{2}:\\d{2}:\\d{2}\\.\\d+\\S+)\\s+caller=(?<source>.+?)\\s+level=(?<severity>\\w+)\\s+(?<log>.*)$",
+					Regex:      "^level=(?<severity>\\w+)\\s+ts=(?<time>\\d{4}-\\d{2}-\\d{2}[Tt].*[zZ])\\s+caller=(?<source>[^\\s]*+)\\s+(?<log>.*)",
 					TimeKey:    "time",
-					TimeFormat: "%Y-%m-%dT%H:%M:%S.%L%z",
+					TimeFormat: "%Y-%m-%dT%H:%M:%S.%L",
 				},
 			},
 		},
